[feat][refactor]Add chat and refactor current code
Implement fully functional ingame chat.
Remove steam integration
Restructure behavior tree assets and logic into separate assets
Separate training and game logic.

+++ Source/BossBattle/Characters/HealthComponent.h
@@ -52,7 +52,7 @@ protected:
 	int MaxHealth = 100;
 
 	UPROPERTY(Replicated, ReplicatedUsing = OnRep_HealthCheck, VisibleInstanceOnly)
-	int Health;
+	float Health;
 
 	bool bDead = false;
